requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
reportlab==4.0.7
PyPDF2==3.0.1
Pillow==10.0.0
sumsub-sdk==0.2.3
